requirements.txt
Django==2.0.5
django-bootstrap3==10.0.1
django-filter==1.1.0
django-query-parameters==0.2.3
django-rest-framework==0.1.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.1.0
pytz==2018.4
sorl-thumbnail==12.4.1
